Reorganize project into hexagonal structure
