+++ src/main/java/com/fontolan/spring/securitykeyvault/example/adapters/security/jwt/JwtUtil.java
@@ -1,0 +1,49 @@
+package com.fontolan.spring.securitykeyvault.example.adapters.security.jwt;
+
+import io.jsonwebtoken.Jwts;
+import io.jsonwebtoken.SignatureAlgorithm;
+import io.jsonwebtoken.io.Decoders;
+import io.jsonwebtoken.security.Keys;
+import org.springframework.beans.factory.annotation.Value;
+import org.springframework.stereotype.Component;
+
+import java.security.Key;
+import java.util.Date;
+
+@Component
+public class JwtUtil {
+    private final Key key;
+    private static final long EXPIRATION_MS = 3600_000; // 1 hour
+
+    public JwtUtil(@Value("${JWT_SECRET:YXdzZW9tZWhzZWNyZXQxMjM0NQ==}") String secretBase64) {
+        byte[] secret = Decoders.BASE64.decode(secretBase64);
+        this.key = Keys.hmacShaKeyFor(secret);
+    }
+
+    public String generateToken(String username) {
+        Date now = new Date();
+        Date exp = new Date(now.getTime() + EXPIRATION_MS);
+        return Jwts.builder()
+                .setSubject(username)
+                .setIssuedAt(now)
+                .setExpiration(exp)
+                .signWith(key, SignatureAlgorithm.HS256)
+                .compact();
+    }
+
+    public String getUsername(String token) {
+        return Jwts.parserBuilder().setSigningKey(key).build()
+                .parseClaimsJws(token)
+                .getBody()
+                .getSubject();
+    }
+
+    public boolean isExpired(String token) {
+        Date expiration = Jwts.parserBuilder().setSigningKey(key).build()
+                .parseClaimsJws(token)
+                .getBody()
+                .getExpiration();
+        return expiration.before(new Date());
+    }
+}
+
