--- src/main/java/com/fontolan/spring/securitykeyvault/example/auth/AuthProvider.java
@@ -1,8 +1,0 @@
-package com.fontolan.spring.securitykeyvault.example.auth;
-
-public enum AuthProvider {
-    LOCAL,
-    GOOGLE,
-    FACEBOOK,
-    KEYCLOAK
-}
