--- src/main/java/com/fontolan/spring/securitykeyvault/example/auth/Role.java
@@ -1,6 +1,0 @@
-package com.fontolan.spring.securitykeyvault.example.auth;
-
-public enum Role {
-    ROLE_USER,
-    ROLE_ADMIN
-}
