--- src/main/java/com/fontolan/spring/securitykeyvault/example/auth/jwt/JwtAuthenticationFilter.java
@@ -1,52 +1,0 @@
-package com.fontolan.spring.securitykeyvault.example.auth.jwt;
-
-import com.fontolan.spring.securitykeyvault.example.auth.tokens.TokenService;
-import com.fontolan.spring.securitykeyvault.example.auth.UserService;
-import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
-import org.springframework.security.core.context.SecurityContextHolder;
-import org.springframework.security.core.userdetails.UserDetails;
-import org.springframework.security.web.authentication.WebAuthenticationDetailsSource;
-import org.springframework.stereotype.Component;
-import org.springframework.web.filter.OncePerRequestFilter;
-
-import jakarta.servlet.FilterChain;
-import jakarta.servlet.ServletException;
-import jakarta.servlet.http.HttpServletRequest;
-import jakarta.servlet.http.HttpServletResponse;
-import java.io.IOException;
-
-@Component
-public class JwtAuthenticationFilter extends OncePerRequestFilter {
-    private final JwtUtil jwtUtil;
-    private final TokenService tokenService;
-    private final UserService userService;
-
-    public JwtAuthenticationFilter(JwtUtil jwtUtil, TokenService tokenService, UserService userService) {
-        this.jwtUtil = jwtUtil;
-        this.tokenService = tokenService;
-        this.userService = userService;
-    }
-
-    @Override
-    protected void doFilterInternal(HttpServletRequest request,
-                                    HttpServletResponse response,
-                                    FilterChain filterChain) throws ServletException, IOException {
-        String authHeader = request.getHeader("Authorization");
-        if (authHeader != null && authHeader.startsWith("Bearer ")) {
-            String token = authHeader.substring(7);
-            try {
-                String username = jwtUtil.getUsername(token);
-                if (!jwtUtil.isExpired(token) && tokenService.isTokenValid(token) &&
-                        SecurityContextHolder.getContext().getAuthentication() == null) {
-                    UserDetails userDetails = userService.loadUserByUsername(username);
-                    UsernamePasswordAuthenticationToken authentication =
-                            new UsernamePasswordAuthenticationToken(userDetails, null, userDetails.getAuthorities());
-                    authentication.setDetails(new WebAuthenticationDetailsSource().buildDetails(request));
-                    SecurityContextHolder.getContext().setAuthentication(authentication);
-                }
-            } catch (Exception ignored) {
-            }
-        }
-        filterChain.doFilter(request, response);
-    }
-}
