--- src/main/java/com/fontolan/spring/securitykeyvault/example/auth/tokens/TokenRepository.java
@@ -1,13 +1,0 @@
-package com.fontolan.spring.securitykeyvault.example.auth.tokens;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-import java.util.Optional;
-import java.util.List;
-
-@Repository
-public interface TokenRepository extends JpaRepository<Token, Long> {
-    Optional<Token> findByToken(String token);
-    List<Token> findByUsernameAndRevokedFalse(String username);
-}
