--- src/main/java/com/fontolan/spring/securitykeyvault/example/controllers/ProductController.java
@@ -1,34 +1,0 @@
-package com.fontolan.spring.securitykeyvault.example.controllers;
-
-import com.fontolan.spring.securitykeyvault.example.entities.Product;
-import com.fontolan.spring.securitykeyvault.example.services.ProductService;
-import org.springframework.data.domain.Page;
-import org.springframework.data.domain.PageRequest;
-import org.springframework.data.domain.Pageable;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RequestParam;
-import org.springframework.web.bind.annotation.RestController;
-import jakarta.servlet.http.HttpServletRequest;
-import org.springframework.security.access.prepost.PreAuthorize;
-
-@RestController
-@RequestMapping("/api/v1/products")
-public class ProductController {
-    private final ProductService productService;
-
-    public ProductController(ProductService productService) {
-        this.productService = productService;
-    }
-
-    @GetMapping
-    @PreAuthorize("hasAnyRole('USER','ADMIN') and @userDeviceService.isTrustedDevice(authentication.name, #request)")
-    public Page<Product> listarProdutos(
-            @RequestParam(defaultValue = "0") int page,
-            @RequestParam(defaultValue = "10") int size,
-            HttpServletRequest request) {
-        Pageable pageable = PageRequest.of(page, size);
-
-        return productService.getAll(pageable);
-    }
-}
