+++ src/main/java/com/fontolan/spring/securitykeyvault/example/domain/model/AuthProvider.java
@@ -1,0 +1,8 @@
+package com.fontolan.spring.securitykeyvault.example.domain.model;
+
+public enum AuthProvider {
+    LOCAL,
+    GOOGLE,
+    FACEBOOK,
+    KEYCLOAK
+}
