+++ src/main/java/com/fontolan/spring/securitykeyvault/example/domain/model/Product.java
@@ -1,0 +1,15 @@
+package com.fontolan.spring.securitykeyvault.example.domain.model;
+
+import jakarta.persistence.*;
+import lombok.Data;
+
+@Data
+@Entity
+@Table(name = "products")
+public class Product {
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Long id;
+    private String name;
+    private Double price;
+}
