+++ src/main/java/com/fontolan/spring/securitykeyvault/example/domain/model/Role.java
@@ -1,0 +1,6 @@
+package com.fontolan.spring.securitykeyvault.example.domain.model;
+
+public enum Role {
+    ROLE_USER,
+    ROLE_ADMIN
+}
