+++ src/main/java/com/fontolan/spring/securitykeyvault/example/domain/repository/ProductRepository.java
@@ -1,0 +1,9 @@
+package com.fontolan.spring.securitykeyvault.example.domain.repository;
+
+import com.fontolan.spring.securitykeyvault.example.domain.model.Product;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+@Repository
+public interface ProductRepository extends JpaRepository<Product, Long> {
+}
