--- src/main/java/com/fontolan/spring/securitykeyvault/example/entities/Product.java
@@ -1,15 +1,0 @@
-package com.fontolan.spring.securitykeyvault.example.entities;
-
-import jakarta.persistence.*;
-import lombok.Data;
-
-@Data
-@Entity
-@Table(name = "products")
-public class Product {
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-    private String name;
-    private Double price;
-}
