+++ src/main/java/com/fontolan/spring/securitykeyvault/example/infrastructure/config/TwoFactorConfig.java
@@ -1,0 +1,14 @@
+package com.fontolan.spring.securitykeyvault.example.infrastructure.config;
+
+import com.warrenstrange.googleauth.GoogleAuthenticator;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+
+@Configuration
+public class TwoFactorConfig {
+    @Bean
+    public GoogleAuthenticator googleAuthenticator() {
+        return new GoogleAuthenticator();
+    }
+}
+
